upnp: make released mappings available for future use
When a client (controller) releases a mapping, the expected behaviour is
that the client deletes it from its local mapping list (the controller
list) and the context makes it available for future use (by the same
controller or any other controller). Currently, when the client releases
a mapping, the context requests a mapping removal from the IGD and
removes it from its local list (the context list).

GitLab: #39
Change-Id: Ic7d54551372f2af9752c1469dc3442db2503a0fb