config: replace jami.net servers by sfl.io
We was using jami.net tun server and bootstrap node in docs, config and
even source code as default. To split concerns, DHTNet now have it's
own turn.sfl.io and bootstrap.sfl.io services for demonstration and
public testing purpose. They are still not intended for production use.

Also change default conf to allow anonymous by default.
This is not a security concern as by default we only allow SSH access,
which is secure and restricted by key / password.

Change-Id: Ibcd3607dcd7f46ed21c0ac396fde0459edc2c92e