upnp: set the status of auto-update mapping to pending when the IGD becomes invalid
Currently, if the IGD becomes invalid, we set the mapping status to FAILED and call unregisterMappings.
unregisterMappings will request another mapping of the same type if the mapping has auto-update enabled.

This patch suggest keeping the same mapping and changing its status to PENDING so it can be requested later,
instead of requesting a new mapping.

Change-Id: Ib34f43afd434f3116e9255a9288874fe0e6f2abd