nat_pmp: remove duplicate logic
This patch removes the custom retry limit and timeout in the
NatPmp::readResponse function; they were not necessary given that
libnatpmp already implements this functionality.

This change means that readResponse can no longer return the
NATPMP_TRYAGAIN error code. The fact that it could previously seems to
be what enabled the bug described in the following issue:
https://git.jami.net/savoirfairelinux/jami-client-qt/-/issues/1898
More precisely, it seems that the readResponse function was repeatedly
returning NATPMP_TRYAGAIN (even though the mapping requests had actually
succeeded), which isn't a "fatal" error (as determined by the
NatPmp::isErrorFatal function) and therefore didn't cause the error
counter to be incremented and the IGD to be invalidated.

GitLab: #43
Change-Id: I1374ba1b40e867399dab0f2dc4f6f75424f4a6b6