UPnPContext: improve port availability check and logging
This makes it impossible for getAvailablePortNumber to fail unless there are literally no available ports, in which case it returns 0. The initial port is randomly generated and will immediately work in most cases. If not, it will be incremented and tested again until reaching the upper bound. Then, it will be set to the lower bound and tested until reaching the initial port - 1.

Change-Id: Ie673e2d7ceb47427006f022ca68423b433533ac9