tests/ice: pass correct argument to getPublicAddress
In IceTest::setUp, we wait for the DHT's public address to be available
because it is needed by several tests. However, this doesn't always work
as expected currently because the tests always use the IPv4 address,
whereas the setUp() function waits for either the IPv4 or the IPv6
address to become available.

Change-Id: I82efc8cf7032fb668b531d4e5deebf481a334301