ChannelSocket: close on read error or EOF signal
Allow the read callback to return zero for a graceful shutdown,
or a negative error code for an error.

In the current implementation, both are handled the same way
because the multiplex layer currently has no error signaling mechanism.

In both cases the channel is shutdown immediately and synchronously,
guaranteeing that no further channel callback is called,
except the shutdown callback.

Change-Id: If9cf4622b7f521558f3b067dcc69d65b05d19c4f